requirements.txt
amqp==5.0.6
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.10.0
django-ipware==4.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.4.0
identify==2.4.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.2
MarkupSafe==2.0.1
nodeenv==1.6.0
openapi-codec==1.3.2
platformdirs==2.4.0
pre-commit==2.15.0
prompt-toolkit==3.0.23
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
wrapt==1.13.3
